+++ communication/src/main/java/com/appcoins/communication/MessageProcessorActivity.java
@@ -5,7 +5,7 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.os.Parcelable;
 
-public abstract class MessageReceiverActivity extends Activity {
+public abstract class MessageProcessorActivity extends Activity {
 
   @Override protected void onCreate(Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);
